Refactor: Improve Readability, Modularity, and Logging
- Restructured FastAPI application for enhanced readability and organization.
- Reorganized application into modules for better modularity.
- Separated configuration settings into 'config.py'.
- Created a router in 'routes.py' and integrated it with the main app.

- Introduce Metrics Class:
  - Implemented Metrics class to manage metric-related queries.
  - Refactored metric endpoints to utilize the Metrics methods.

- Implement CSV Processor Class:
  - Introduced CSVProcessor class to handle CSV processing.
  - Moved CSV processing logic from the main app to this class.

- Enhance Database Handling:
  - Created 'database.py' module for database-related functions.
  - Transferred BigQuery schema and insertion logic to this module.

- Configure Logging:
  - Setup 'logger.py' module for streamlined logger configuration.
  - Defined logging format and added a stream handler.
